Add files via upload
This is the final Model of emotion detection. You can download the project and test file and later import the file to test and add the weights and run the test file and BOOM, your real time emotion detection project is done. 

Let me know if you require the training code , make a pull request and I will provide it personally